# server/docs/openAPI.yml
openapi: 3.0.3
info:
  title: User API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /user/register:
    post:
      summary: Register a new user
      operationId: registerUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Validation error — malformed or missing fields
          content:
            text/plain:
              schema:
                type: string
              examples:
                invalid_email:
                  summary: Invalid email format
                  value: "invalid email format"
                parsing_error:
                  summary: Invalid date
                  value: 'code=400, message=Bad Request, err=parsing time "190Z" as "2006-01-02T15:04:05Z07:00": cannot parse "190Z" as "2006"'
                validation_error:
                  summary: Name below minimum length
                  value: "Key: 'RegisterRequest.Email' Error:Field validation for 'Email' failed on the 'email' tag\nKey: 'RegisterRequest.Password' Error:Field validation for 'Password' failed on the 'required' tag\nKey: 'RegisterRequest.Name' Error:Field validation for 'Name' failed on the 'min' tag\nKey: 'RegisterRequest.Surname' Error:Field validation for 'Surname' failed on the 'required' tag"

  /user/login:
    post:
      summary: Login the user
      operationId: loginUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error — malformed or missing fields
          content:
            text/plain:
              schema:
                type: string
              examples:
                user_not_found:
                  summary: User does not exist
                  value: "User not found"
                validation_error:
                  summary: Name below minimum length
                  value: "Key: 'RegisterRequest.Email' Error:Field validation for 'Email' failed on the 'email' tag\nKey: 'RegisterRequest.Password' Error:Field validation for 'Password' failed on the 'required' tag\nKey: 'RegisterRequest.Name' Error:Field validation for 'Name' failed on the 'min' tag\nKey: 'RegisterRequest.Surname' Error:Field validation for 'Surname' failed on the 'required' tag"

  /user/:
    get:
      summary: Get the logged in user details
      operationId: userDetails
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetailsResponse"
        "404":
          description: No user with id that is inside the token
          content:
            text/plain:
              schema:
                type: string
              examples:
                user_not_found:
                  summary: User does not exist
                  value: "User not found"
        "401":
          description: Invalid token
          content:
            text/plain:
              schema:
                type: string
              example: "Malformed token"

    delete:
      summary: Delete the logged in user
      operationId: deleteUser
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User deleted successfully
          content:
            text/plain:
              schema:
                type: string
              example: "User deleted successfully"
        "404":
          description: No user with id that is inside the token
          content:
            text/plain:
              schema:
                type: string
              examples:
                user_not_found:
                  summary: User does not exist
                  value: "User not found"
        "401":
          description: Invalid token
          content:
            text/plain:
              schema:
                type: string
              example: "Malformed token"

    put:
      summary: Modify user data
      operationId: UpdateUser
      tags:
        - User
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequest"
      responses:
        "200":
          description: User deleted successfully
          content:
            text/plain:
              schema:
                type: string
              example: "User deleted successfully"
        "404":
          description: No user with id that is inside the token
          content:
            text/plain:
              schema:
                type: string
              examples:
                user_not_found:
                  summary: User does not exist
                  value: "User not found"
        "401":
          description: Invalid token
          content:
            text/plain:
              schema:
                type: string
              example: "Malformed token"
  /user/all:
    get:
      summary: Get all users
      operationId: getAllUsers
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User deleted successfully
          content:
            text/plain:
              schema:
                type: string
              example: "User deleted successfully"
        "404":
          description: No user with id that is inside the token
          content:
            text/plain:
              schema:
                type: string
              examples:
                user_not_found:
                  summary: User does not exist
                  value: "User not found"
        "401":
          description: Invalid token
          content:
            text/plain:
              schema:
                type: string
              example: "Malformed token"

  /chats/:
    get:
      summary: Get the users chats
      operationId: chatsGet
      tags:
        - Chats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetChatsResponse"
        "404":
          description: No user with id that is inside the token
          content:
            text/plain:
              schema:
                type: string
              examples:
                user_not_found:
                  summary: User does not exist
                  value: "User not found"
        "401":
          description: Invalid token
          content:
            text/plain:
              schema:
                type: string
              example: "Malformed token"
    post:
      summary: Add a chat
      operationId: chatsAd
      tags:
        - Chats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetChatsResponse"
        "404":
          description: No user with id that is inside the token
          content:
            text/plain:
              schema:
                type: string
              examples:
                user_not_found:
                  summary: User does not exist
                  value: "User not found"
        "401":
          description: Invalid token
          content:
            text/plain:
              schema:
                type: string
              example: "Malformed token"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
        - surname
        - birthdate
      properties:
        email:
          type: string
          format: email
          description: Valid email address
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 32
          description: Password between 8 and 32 characters
          example: securePass1
        name:
          type: string
          minLength: 3
          maxLength: 60
          description: User's first name
          example: John
        surname:
          type: string
          minLength: 3
          maxLength: 60
          description: User's last name
          example: Doe
        birthdate:
          type: string
          format: date-time
          description: Date of birth in RFC 3339 / ISO 8601 format
          example: "1990-06-15T00:00:00Z"

    RegisterResponse:
      type: object
      properties:
        token:
          type: string
          description: Signed JWT for authenticating subsequent requests
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxMCwiZXhwIjoxNzgwNTE0MDQxLCJpYXQiOjE3ODA0Mjc2NDF9.yQ1rl4Mkxwvlf5x2yStqy-iVRtkj4NcTNEJ_brhxhRU
        user_id:
          type: integer
          format: uint
          description: Unique ID of the newly created user
          example: 42

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Valid email address
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 32
          description: Password between 8 and 32 characters
          example: securePass1

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Signed JWT for authenticating subsequent requests
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxMCwiZXhwIjoxNzgwNTE0MDQxLCJpYXQiOjE3ODA0Mjc2NDF9.yQ1rl4Mkxwvlf5x2yStqy-iVRtkj4NcTNEJ_brhxhRU
        user_id:
          type: integer
          format: uint
          description: Unique ID of the newly created user
          example: 42

    DetailsResponse:
      type: object
      properties:
        id:
          type: integer
          format: uint
          description: Unique ID of the user
          example: 42
        email:
          type: string
          description: Users email
          example: john.doe@example.com
        name:
          type: string
          description: Users first name
          example: John
        surname:
          type: string
          description: Users last name
          example: Doe
        birthdate:
          type: string
          format: date-time
          description: Users birthdate as a timestamp

    UpdateRequest:
      type: object
      properties:
        email:
          type: string
          description: Users email
          example: john2.doe@example.com
        name:
          type: string
          description: Users first name
          example: John
        surname:
          type: string
          description: Users last name
          example: Doe
        birthdate:
          type: string
          format: date-time
          description: Users birthdate as a timestamp

    ChatMessage:
      properties:
        id:
          type: string
          description: Unique ID of the message
          example: "afcabefbbafacaebf243"
        author:
          type: string
          description: Unique ID of the user that sent the message
          example: "afcabefbbafacaebf243"
        message:
          type: string
          description: The message content
          example: "Hello, how are you?"

    Chat:
      properties:
        id:
          type: string
          description: Unique ID of the chat
          example: "afcabefbbafacaebf243"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        members:
          type: array
          items:
            type: string
            description: Unique ID of the user that is a member of the chat
            example: "afcabefbbafacaebf243"

    GetChatsResponse:
      type: object
      properties:
        chats:
          type: array
          items:
            $ref: "#/components/schemas/Chat"